Use IOnLanguageServerStarted
